chore(array-utils): convert "gifts" into classes
