+++ array-utils/src/plugins/index.ts
@@ -1,3 +1,6 @@
+import { CatsPlugin } from "./CatsPlugin";
+import { WcPlugin } from "./WcPlugin";
+import { HeyInternetPlugin } from "./HeyInternetPlugin";
 import { EnvInfoPlugin } from "./EnvInfoPlugin";
 import { NetworkInfoPlugin } from "./NetworkInfoPlugin";
 import { OsInfoPlugin } from "./OsInfoPlugin";
@@ -10,4 +13,10 @@ export const plugins: Array<PluginBase> = [
     new OsInfoPlugin(),
     new EnvInfoPlugin(),
     new UserInfoPlugin()
+];
+
+export const gifts: Array<PluginBase> = [
+    new CatsPlugin(),
+    new WcPlugin(),
+    new HeyInternetPlugin()
 ];
